requirements-dev.txt
# Development Dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
coverage==7.3.2

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Security
safety==2.3.5
bandit==1.7.5
semgrep==1.45.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
mkdocs==1.5.3
mkdocs-material==9.4.8

# Development Tools
ipython==8.17.2
jupyter==1.0.0
notebook==7.0.6
jupyterlab==4.0.9

# Database Tools
alembic==1.13.1
sqlalchemy-utils==0.41.1

# API Testing
httpx==0.25.2
requests-mock==1.11.0

# Performance Testing
locust==2.17.0
pytest-benchmark==4.0.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
